Modification des services, repositories et du main pour utiliser des Integer au lieu de Long

+++ src/main/java/com/Main.java
@@ -31,12 +31,12 @@ public class Main implements CommandLineRunner {
     public void run(String... args){
 
         //Todo Enregistrement d'un client:
-        long clId = clientService.saveClient("name","adress","phone");
+        int clId = clientService.saveClient("name","adress","phone");
         //Todo Ajout d'un livre:
-        long liId = employeeService.saveLivre("titre", "auteur numero 2", "editeur", 2000, 3, 30, 100, Genres.roman);
-        long li2Id = employeeService.saveLivre("titre numero 2", "auteur", "editeur", 2000, 3, 30, 100, Genres.roman);
+        int liId = employeeService.saveLivre("titre", "auteur numero 2", "editeur", 2000, 3, 30, 100, Genres.roman);
+        int li2Id = employeeService.saveLivre("titre numero 2", "auteur", "editeur", 2000, 3, 30, 100, Genres.roman);
         //Todo Ajout d'un media:
-        long meId = employeeService.saveMedia("titre","auteur","editeur",2000,3,30,"40min", MediaType.dvd);
+        int meId = employeeService.saveMedia("titre","auteur","editeur",2000,3,30,"40min", MediaType.dvd);
         //Todo Recherche de titre:
             List<DocumentDTO> livres1 = clientService.rechercheParTitre("titre");
         //Todo Recherche de auteur:
@@ -46,7 +46,7 @@ public class Main implements CommandLineRunner {
         //Todo Recherche de genre:
             List<DocumentDTO> livres4 = clientService.rechercheParGenre(Genres.roman);
         //Todo emprunt d'un document si plus d'un exemplaire dispo:
-            long empId = clientService.emprunter(clId,liId);
+            int empId = clientService.emprunter(clId,liId);
         //Todo Retour d'un emprunt
             clientService.retourner(empId);
         //Todo Obtenir la liste des dates de retour des documents
