+++ src/main/java/com/repository/DetteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DetteRepository extends JpaRepository<Dette,Long> {
+public interface DetteRepository extends JpaRepository<Dette,Integer> {
 }
